consolidate hash dispatch functions
