+++ dask/dataframe/backends.py
@@ -1,6 +1,6 @@
 from .methods import concat_dispatch
 from .core import get_parallel_type, meta_nonempty, make_meta
-from .utils import hash_object_dispatch
+from .utils import hash_object_dispatch, group_split_dispatch
 
 
 ######################################
@@ -10,6 +10,7 @@ from .utils import hash_object_dispatch
 
 @concat_dispatch.register_lazy("cudf")
 @hash_object_dispatch.register_lazy("cudf")
+@group_split_dispatch.register_lazy("cudf")
 @get_parallel_type.register_lazy("cudf")
 @meta_nonempty.register_lazy("cudf")
 @make_meta.register_lazy("cudf")
