+++ dask/dataframe/utils.py
@@ -462,6 +462,18 @@ def hash_object_pandas(
     )
 
 
+group_split_dispatch = Dispatch("group_split_dispatch")
+
+
+@group_split_dispatch.register((pd.DataFrame, pd.Series, pd.Index))
+def group_split_pandas(df, c, k):
+    indexer, locations = pd._libs.algos.groupsort_indexer(c, k)
+    df2 = df.take(indexer)
+    locations = locations.cumsum()
+    parts = [df2.iloc[a:b] for a, b in zip(locations[:-1], locations[1:])]
+    return dict(zip(range(k), parts))
+
+
 _simple_fake_mapping = {
     "b": np.bool_(True),
     "V": np.void(b" "),
